pyproject.toml
[project]
name = "docqa"
version = "0.1.0"
description = "DocQA: Document Q&A using LangChain and Google GenAI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.16",
    "langchain-community>=0.2.16",
    "langchain-google-genai>=1.0.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.38.0",
    "pypdf>=4.3.1",
    "faiss-cpu>=1.8.0.post1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.3",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "SIM",  # flake8-simplify
    "I",  # isort
]

[tool.ruff.format]
docstring-code-format = true
